+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
 * FEATURE: export work details as XLSX
 * FEATURE: resizable table columns (using CKEditoR)
+* FEATURE: highlight text and make comments
 * Make it easier to override colophons
 * Rename output renderers to exporters, so as not to clash with DRF renderers